+++ src/controllers/DefaultController.php
@@ -1,23 +1,38 @@
 <?php
-namespace mailchimp\controllers;
+
+namespace luya\mailchimp\controllers;
 
 use Yii;
 use luya\base\DynamicModel;
 use yii\base\InvalidConfigException;
 use luya\Exception;
+use luya\web\Controller;
 use \Mailchimp;
+use luya\mailchimp\helpers\MailchimpHelper;
 
-class DefaultController extends \luya\web\Controller
+class DefaultController extends Controller
 {
     /**
-     * @var null|bool if null no status information has been assigned, if false a global error happend (could not send mail), if true
-     * the form has been sent successfull.
+     * Initializer
+     * @todo use getter exception inside module attribute instead of initializer of controller.
      */
-    public $success = null;
-
-    const ERROR_EMAIL_ALREADY_SUBSCRIBED = 214;
-
-
+    public function init()
+    {
+        parent::init();
+        
+        if ($this->module->listId === null) {
+            throw new Exception("The MailChimp list Id must be defined.");
+        }
+        
+        if ($this->module->mailchimpApi === null) {
+            throw new Exception("The MailChimp API key must be defined.");
+        }
+        
+        if ($this->module->attributes === null) {
+            throw new Exception("The attributes attributed must be defined with an array of available attributes.");
+        }
+    }
+    
     /**
      * @param $model
      * @param $alias name of the mailchimp group in our dynamic model
@@ -31,9 +46,31 @@ class DefaultController extends \luya\web\Controller
                 $returnArray[] = $element;
             }
         }
+        
         return $returnArray;
     }
 
+    /**
+     * 
+     * @throws InvalidConfigException
+     * @return \luya\base\DynamicModel
+     */
+    private function generateModelFromModule()
+    {
+        $model = new DynamicModel($this->module->attributes);
+        $model->attributeLabels = $this->module->attributeLabels;
+        
+        foreach ($this->module->rules as $rule) {
+            if (is_array($rule) && isset($rule[0], $rule[1])) {
+                $model->addRule($rule[0], $rule[1], isset($rule[2]) ? $rule[2] : []);
+            } else {
+                throw new InvalidConfigException('Invalid validation rule: a rule must specify both attribute names and validator type.');
+            }
+        }
+        
+        return $model;
+    }
+    
     /**
      * Index Action
      *
@@ -42,37 +79,23 @@ class DefaultController extends \luya\web\Controller
      */
     public function actionIndex()
     {
-        // create dynamic model
-        $model = new DynamicModel($this->module->attributes);
-        $model->attributeLabels = $this->module->attributeLabels;
-        $error = null;
-
-        foreach ($this->module->rules as $rule) {
-            if (is_array($rule) && isset($rule[0], $rule[1])) {
-                $model->addRule($rule[0], $rule[1], isset($rule[2]) ? $rule[2] : []);
-            } else {
-                throw new InvalidConfigException('Invalid validation rule: a rule must specify both attribute names and validator type.');
-            }
-        }
+        $model = $this->generateModelFromModule();
 
         if ($model->load(Yii::$app->request->post()) && $model->validate()) {
-
             if ((intval(time()) - intval(Yii::$app->session->get('renderTime', 0))) < $this->module->spamDetectionDelay) {
                 throw new Exception("We haved catched a spam contact form with the values: " . print_r($model->attributes, true));
             }
 
-            $mailchimp = new Mailchimp($this->module->mailchimpApi);
-
             $merge_vars = null;
-            foreach ($model->attributes as $key=>$value) {
-                if ($key != 'email') {
+            foreach ($model->attributes as $key => $value) {
+                if ($key != $this->module->attributeEmailField) {
                     $merge_vars[$key] = $value;
                 }
             }
 
             // add interest groups
             foreach ($this->module->groups as $group) {
-                $merge_vars['groupings'] = [
+                $merge_vars['groupings'][] = [
                     [
                         'id' => $group['id'],
                         'groups' => $this->getGroupAttributes($model, $group['alias'])
@@ -81,51 +104,34 @@ class DefaultController extends \luya\web\Controller
             }
 
             try {
-                $mailchimp->lists->subscribe($this->module->listId, array('email' => $model->email), $merge_vars, false, false, false, false);
-            } catch(\Mailchimp_Error $e) {
-                $error = $e;
+                (new MailchimpHelper($this->module->mailchimpApi))->subscribe($this->module->listId, $model->{$this->module->attributeEmailField}, $merge_vars);
+            } catch (\Mailchimp_Error $e) {
+                $model->addError($this->module->attributeEmailField, $e->getMessage());
             }
-
-            if (empty($error)) {
-                // send admin success mail
-                if ($this->module->recipients !== null) {
+            
+            if (!$model->hasErrors()) {
+                if ($this->module->recipients) {
                     $mail = Yii::$app->mail->compose('['.Yii::$app->siteTitle.'] newsletter registration', $this->renderPartial('_mail', ['model' => $model]));
                     $mail->adresses($this->module->recipients);
-
                     if ($mail->send()) {
-                        $this->success = true;
-
                         // callback
                         $cb = $this->module->callback;
                         if (is_callable($cb)) {
                             $cb($model);
                         }
-
-                    } else {
-                        $this->success = false;
                     }
-                } else {
-                    $this->success = true;
                 }
+                
+                Yii::$app->session->setFlash('mailchimpSuccess');
+                
+                return $this->refresh();
             }
         }
 
         Yii::$app->session->set('renderTime', time());
-/*
-        // compile groups
-        $groups = null;
-        foreach ($this->module->groups as $group) {
-            
-            foreach ($group->array as $element) {
-                $groups
-            }
-        }
-        */
         
         return $this->render('index', [
             'model' => $model,
-            "error" => $error,
-            'success' => $this->success,
         ]);
     }
 }