+++ CHANGELOG.md
@@ -3,8 +3,13 @@ LUYA MODULE MAILCHIMP CHANGELOG
 
 All notable changes to this project will be documented in this file. This project make usage of the [Yii Versioning Strategy](https://github.com/yiisoft/yii2/blob/master/docs/internals/versions.md).
 
-1.0.1 (in progress)
+1.1.0 (in progress)
 -------------------
+
++ [#5](https://github.com/luyadev/luya-module-mailchimp/issues/5) Added Helper method in order to make API calls without instantiating the Mailchimp Module.
++ [#4](https://github.com/luyadev/luya-module-mailchimp/issues/4) Changed the application structure and namespaces of the application.
++ [#3](https://github.com/luyadev/luya-module-mailchimp/issues/3) Removed error and success variables and moved to standard active form validation with response redirect and flash message.
++ Added LUYA Testsuite in order to make basic application tests.
 
 1.0.0 (25. April 2017)
 ----------------------